fix warning C4777: '_snprintf_s' : format string '%d' requires an argument of type 'int', but variadic argument 1 has type 'DWORD' (compiling source file ../../thirdparty/owemdjee/OpenEXR/src/lib/OpenEXRCore/context.c)

+++ src/lib/OpenEXRCore/internal_win32_file_impl.h
@@ -411,7 +411,7 @@ make_temp_filename (exr_context_t ret)
     uint64_t    tlen, newlen;
     const char* srcfile = ret->filename.str;
     int         nwr =
-        _snprintf_s (tmproot, 32, _TRUNCATE, "tmp.%d", GetCurrentProcessId ());
+        _snprintf_s (tmproot, 32, _TRUNCATE, "tmp.%u", (unsigned int)GetCurrentProcessId ());
     if (nwr >= 32)
         return ret->report_error (
             ret,
